Minor refactor and renaming
