+++ iOS/Settings/Appearance View/DisplayAndBehaviorsView.swift
@@ -15,13 +15,7 @@ struct DisplayAndBehaviorsView: View {
 	var body: some View {
 		List {
 			Section("Application") {
-				HStack {
-					appLightButton()
-					Spacer()
-					appDarkButton()
-					Spacer()
-					appAutomaticButton()
-				}
+				ColorPaletteSelectorView()
 				.listRowBackground(Color.clear)
 			}
 			
@@ -44,62 +38,7 @@ struct DisplayAndBehaviorsView: View {
 		.tint(Color(uiColor: AppAssets.primaryAccentColor))
     }
 	
-	func appLightButton() -> some View {
-		VStack(spacing: 4) {
-			Image("app.appearance.light")
-				.resizable()
-				.aspectRatio(contentMode: .fit)
-				.frame(width: 40.0, height: 40.0)
-			Text("Always Light")
-				.font(.subheadline)
-			if AppDefaults.userInterfaceColorPalette == .light {
-				Image(systemName: "checkmark.circle.fill")
-					.foregroundColor(Color(uiColor: AppAssets.primaryAccentColor))
-			} else {
-				Image(systemName: "circle")
-			}
-		}.onTapGesture {
-			AppDefaults.userInterfaceColorPalette = .light
-		}
-	}
 	
-	func appDarkButton() -> some View {
-		VStack(spacing: 4) {
-			Image("app.appearance.dark")
-				.resizable()
-				.aspectRatio(contentMode: .fit)
-				.frame(width: 40.0, height: 40.0)
-			Text("Always Dark")
-				.font(.subheadline)
-			if AppDefaults.userInterfaceColorPalette == .dark {
-				Image(systemName: "checkmark.circle.fill")
-					.foregroundColor(Color(uiColor: AppAssets.primaryAccentColor))
-			} else {
-				Image(systemName: "circle")
-			}
-		}.onTapGesture {
-			AppDefaults.userInterfaceColorPalette = .dark
-		}
-	}
-	
-	func appAutomaticButton() -> some View {
-		VStack(spacing: 4) {
-			Image("app.appearance.automatic")
-				.resizable()
-				.aspectRatio(contentMode: .fit)
-				.frame(width: 40.0, height: 40.0)
-			Text("Use System")
-				.font(.subheadline)
-			if AppDefaults.userInterfaceColorPalette == .automatic {
-				Image(systemName: "checkmark.circle.fill")
-					.foregroundColor(Color(uiColor: AppAssets.primaryAccentColor))
-			} else {
-				Image(systemName: "circle")
-			}
-		}.onTapGesture {
-			AppDefaults.userInterfaceColorPalette = .automatic
-		}
-	}
 	
 }
